azure-ai-agent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
azureopenai==0.0.1
certifi==2025.7.14
charset-normalizer==3.4.2
frozenlist==1.7.0
idna==3.10
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
multidict==6.6.3
propcache==0.3.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
urllib3==2.5.0
yarl==1.20.1
